refactor(runtime): simplify delegation by removing the wrapper type
The delegator type introduced in the previous commit was a pure
back-pointer to LocalRuntime with no state of its own — every
method just reached through `d.rt` to the runtime. That extra
hop made every callsite read `r.delegator.runForwarding(...)`
when it could just say `r.runForwarding(...)`.

- Drop the delegator type and the LocalRuntime.delegator field.
  runForwarding and runCollecting move onto *LocalRuntime
  directly, matching the existing pattern (handleTaskTransfer,
  handleRunSkill, RunAgent are all already runtime methods).
- Merge delegator.go back into agent_delegation.go now that
  there is no "delegator" inside it. Single file = single
  feature: types, helpers, run methods, and tool handlers all
  live together (~440 lines).
- Add a small swapCurrentAgent(from, to, evts) helper that
  performs the AgentSwitching/AgentInfo emission and
  setCurrentAgent in one place and returns a restore closure.
  runForwarding now expresses the agent-swap as a single
  `defer r.swapCurrentAgent(callerAgent, child, evts)()` line,
  replacing the two split `if req.SwitchCurrentAgent` blocks.
- As a side effect, the AgentSwitching(true) event now fires
  AFTER the child agent is resolved instead of before, so a
  failed child lookup no longer emits a stray "switching to X"
  event followed by a "switching back" event.

Behaviour, span shapes, event ordering of the success path, and
the public API surface (SubSessionConfig and helper functions
covered by agent_delegation_test.go) are unchanged.
